python/requirements.txt
fastapi==0.70.1
uvicorn==0.16.0
py-algorand-sdk==1.8.0
ipfshttpclient==0.7.0
python-dotenv==0.19.2
arweave-python-client==1.0.14
reach-rpc-client==0.1.2.202107081228
redis==4.0.2
wrapt-timeout-decorator==1.3.1
pycryptodome==3.12.0
pyairtable==1.0.0.post1
async-timeout==4.0.2